Fix: Fraud mangles stack on C prim calls.
The Fraud compiler didn't have any test cases that mix the use of
`let` and the C implemented primitives like `write-byte`.  The
compiler was not issuing instructions to fix up the stack pointer
before calling the primitives, which means the environment gets
mangled when non-empty.  Added a test case to illustrate the bug
and fixed it.